implemented TicketPanel

+++ src/main/java/it/unibo/smartcity/model/api/TariffaBiglietto.java
@@ -4,7 +4,7 @@ public interface TariffaBiglietto {
 
     String getNome();
 
-    double getDurata();
+    int getDurata();
 
     double getPrezzo();
